requirements.txt
#requirements
asgiref==3.7.2
attrs==23.2.0   
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0    
channels==4.0.0
constantly==23.10.4
cryptography==42.0.5
daphne==4.0.0
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
et-xmlfile==1.1.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
numpy==1.26.4
openpyxl==3.1.4
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytz==2023.3
redis==5.0.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2023.3
xlrd==2.0.1
zope.interface==6.2